# docs/swagger-openapi.yml
openapi: 3.0.1
info:
  title: Tài liệu REST API Nexora-Commerce
  description: Tài liệu RESTful API hiệu năng cao cho hệ thống thương mại điện tử
    Nexora-Commerce phục vụ tích hợp frontend.
  contact:
    name: Nexora Core Dev Team
    email: dev@nexoracommerce.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Máy chủ phát triển cục bộ (Local Development Server)
security:
- bearerAuth: []
...
paths:
...
  /api/v1/products/{productId}:
    get:
      tags:
      - Sản phẩm
      summary: Lấy chi tiết sản phẩm bằng ID
      description: Truy xuất toàn bộ thông tin chi tiết của một sản phẩm trong danh
        mục dựa theo mã sản phẩm duy nhất.
      operationId: getProductById
      parameters:
      - name: productId
        in: path
        description: Mã sản phẩm duy nhất
        required: true
        schema:
          type: string
        example: P001
      responses:
        '200':
          description: Lấy chi tiết sản phẩm thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseProductResponse'
        '404':
          description: Không tìm thấy sản phẩm dưới ID đã cung cấp
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseProductResponse'
    put:
      tags:
      - Sản phẩm
      summary: Cập nhật thông tin sản phẩm
      description: Yêu cầu quyền ADMIN. Cập nhật các thuộc tính của sản phẩm và tải
        lên/thay đổi hình ảnh sản phẩm.
      operationId: updateProduct
      parameters:
      - name: productId
        in: path
        description: Mã sản phẩm duy nhất
        required: true
        schema:
          type: string
        example: P001
      - name: request
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/ProductFormRequest'
      responses:
        '403':
          description: Không đủ quyền hạn (yêu cầu vai trò ADMIN)
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseProductResponse'
        '400':
          description: Dữ liệu đầu vào không hợp lệ
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseProductResponse'
        '404':
          description: Không tìm thấy sản phẩm
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseProductResponse'
        '401':
          description: Mã xác thực JWT thiếu hoặc không hợp lệ
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseProductResponse'
        '200':
          description: Cập nhật sản phẩm thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseProductResponse'
      security:
      - bearerAuth: []
    delete:
      tags:
      - Sản phẩm
      summary: Xóa một sản phẩm
      description: Yêu cầu quyền ADMIN. Xóa vĩnh viễn sản phẩm khỏi danh mục trong
        cơ sở dữ liệu.
      operationId: deleteProduct
      parameters:
      - name: productId
        in: path
        description: Mã sản phẩm duy nhất
        required: true
        schema:
          type: string
        example: P001
      responses:
        '403':
          description: Không đủ quyền hạn (yêu cầu vai trò ADMIN)
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '404':
          description: Không tìm thấy sản phẩm
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '401':
          description: Mã xác thực JWT thiếu hoặc không hợp lệ
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '200':
          description: Xóa sản phẩm thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
      security:
      - bearerAuth: []
...
  /api/v1/products:
    get:
      tags:
      - Sản phẩm
      summary: Lấy danh sách sản phẩm với bộ lọc và tìm kiếm
      description: Lấy danh sách sản phẩm phân trang, mặc định sắp xếp theo ID tăng
        dần. Hỗ trợ lọc theo danh mục, thương hiệu và tìm kiếm mờ theo từ khóa.
      operationId: getAllProducts
      parameters:
      - name: keyword
        in: query
        required: false
        schema:
          type: string
      - name: categoryId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: brandId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: minPrice
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: maxPrice
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: Lấy danh sách trang sản phẩm thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListProductResponse'
    post:
      tags:
      - Sản phẩm
      summary: Tạo sản phẩm mới
      description: Yêu cầu quyền ADMIN. Khởi tạo một sản phẩm mới trong danh mục cùng
        các tệp hình ảnh tải lên ở định dạng multipart/form-data.
      operationId: createProduct
      parameters:
      - name: request
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/ProductFormRequest'
      responses:
        '403':
          description: Không đủ quyền hạn (yêu cầu vai trò ADMIN)
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseProductResponse'
        '400':
          description: Dữ liệu đầu vào không hợp lệ
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseProductResponse'
        '201':
          description: Tạo sản phẩm thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseProductResponse'
        '401':
          description: Mã xác thực JWT thiếu hoặc không hợp lệ
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseProductResponse'
      security:
      - bearerAuth: []
    delete:
      tags:
      - Sản phẩm
      summary: Xóa hàng loạt sản phẩm (Đã vô hiệu hóa)
      description: Yêu cầu quyền ADMIN. Hành động này bị vô hiệu hóa an toàn và luôn
        trả về lỗi 403 để tránh tai nạn xóa toàn bộ danh mục sản phẩm.
      operationId: deleteAllProducts
      responses:
        '403':
          description: Tính năng xóa hàng loạt đã bị vô hiệu hóa để đảm bảo an toàn
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
      security:
      - bearerAuth: []
...
  /api/v1/products/{productId}/stock:
    get:
      tags:
      - Sản phẩm
      summary: Lấy số lượng tồn khả dụng của sản phẩm
      description: Truy cập trực tiếp số lượng tồn kho khả dụng trong kho của một
        sản phẩm cụ thể.
      operationId: getAvailableStock
      parameters:
      - name: productId
        in: path
        description: Mã sản phẩm duy nhất
        required: true
        schema:
          type: string
        example: P001
      responses:
        '200':
          description: Lấy số lượng tồn kho thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInteger'
        '404':
          description: Không tìm thấy sản phẩm
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInteger'
  /api/v1/products/{productId}/availability:
    get:
      tags:
      - Sản phẩm
      summary: Kiểm tra trạng thái khả dụng của sản phẩm
      description: Kiểm tra xem sản phẩm còn hàng trong kho và có sẵn sàng để bán
        hay không.
      operationId: isProductAvailable
      parameters:
      - name: productId
        in: path
        description: Mã sản phẩm duy nhất
        required: true
        schema:
          type: string
        example: P001
      responses:
        '200':
          description: Kiểm tra trạng thái khả dụng thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBoolean'
        '404':
          description: Không tìm thấy sản phẩm
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBoolean'
...
  /api/v1/admin/products:
    get:
      tags:
      - Quản lý Sản phẩm (Admin)
      summary: Lấy toàn bộ sản phẩm (Admin)
      description: Yêu cầu quyền ADMIN. Lấy danh sách phân trang sản phẩm với các
        bộ lọc dành riêng cho giao diện quản trị viên.
      operationId: getAllProducts_1
      parameters:
      - name: keyword
        in: query
        required: false
        schema:
          type: string
      - name: categoryId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: brandId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: minPrice
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: maxPrice
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: Lấy danh sách sản phẩm thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListProductResponse'
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
...
    ProductFormRequest:
      required:
      - name
      - price
      - status
      type: object
      properties:
        id:
          maxLength: 100
          minLength: 0
          type: string
        name:
          maxLength: 255
          minLength: 1
          type: string
        description:
          maxLength: 1000
          minLength: 0
          type: string
        price:
          type: integer
          format: int64
        status:
          type: string
        image:
          type: string
          format: binary
        variants:
          type: string
    ApiResponseProductResponse:
      type: object
      properties:
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/ProductResponse'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    ProductResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
        status:
          type: string
        imageUrl:
          type: string
        imagePublicId:
          type: string
        variants:
          type: array
          items:
            $ref: '#/components/schemas/ProductVariantResponse'
    ProductVariantAttributeResponse:
      type: object
      properties:
        name:
          type: string
        value:
          type: string
    ProductVariantResponse:
      type: object
      properties:
        sku:
          type: string
        productName:
          type: string
        imageUrl:
          type: string
        price:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
        reservedQuantity:
          type: integer
          format: int32
        soldQuantity:
          type: integer
          format: int32
        attributes:
          type: array
          items:
            $ref: '#/components/schemas/ProductVariantAttributeResponse'
...
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: object
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
...
    ApiResponseListProductResponse:
      type: object
      properties:
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    ApiResponseInteger:
      type: object
      properties:
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: integer
          format: int32
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    ApiResponseBoolean:
      type: object
      properties:
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: boolean
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Please paste your JWT Access Token in the box below to authorize.
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT